+++ src/components/app/default-layout.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import {Route, Switch, Redirect} from 'react-router-dom';
+import { Route, Switch } from 'react-router-dom';
 import DefaultHeader from '../shared/header/defaultheader'
 import Footer from '../shared/footer/footer'
 import Contact from '../contact/contact'
