+++ src/components/app/main-layout.js
@@ -1,17 +1,16 @@
 import React from 'react'
-import {Route, Switch, Redirect} from 'react-router-dom';
+import { Route, Switch } from 'react-router-dom';
 import MainHeader from '../shared/header/mainheader'
 import Footer from '../shared/footer/footer'
 import Main from '../main/main'
 
-//This is the Layout for the main content page
-
+// This is the Layout for the main content page
 const MainLayout = () => {
   return (
     <div>
       <MainHeader />
         <Switch>
-          <Route exact path='/' component={Main} />
+          <Route exact path='/' component={ Main } />
         </Switch>
       <Footer />
     </div>
