+++ src/components/shared/header/navbar.js
@@ -1,8 +1,8 @@
-import React, {useState, useEffect, Fragment} from 'react'
-import {Link} from 'react-router-dom';
+import React, { useState, useEffect, Fragment } from 'react'
+import { Link } from 'react-router-dom';
 import BurgerMenu from '../images/burger-menu.svg'
 
-const NavBar = ({showSectionAnchors}) => {
+const NavBar = ({ showSectionAnchors }) => {
   const [projectAnchor, setProjectAnchor] = useState(null);
   const [skillsAnchor, setSkillsAnchor] = useState(null);
   const [showDropdown, setShowDropdown] = useState(false);
@@ -13,7 +13,7 @@ const NavBar = ({showSectionAnchors}) => {
   }, []);
 
   const handleClick = (anchor) => {
-    window.scrollTo({behavior: 'smooth', block: 'start', top: anchor.offsetTop - 60})
+    window.scrollTo({ behavior: 'smooth', block: 'start', top: anchor.offsetTop - 60 })
     setShowDropdown(false)
   }
 
@@ -24,10 +24,10 @@ const NavBar = ({showSectionAnchors}) => {
           showSectionAnchors ? 
           <Fragment>
             <div className='navbar-link-container'>
-              <div onClick={() => {handleClick(skillsAnchor)}} className='navbar-link'>Skills</div>
+              <div onClick={() => { handleClick(skillsAnchor) }} className='navbar-link'>Skills</div>
             </div>
             <div className='navbar-link-container'>
-              <div onClick={() => {handleClick(projectAnchor)}} className='navbar-link'>Projects</div>
+              <div onClick={() => { handleClick(projectAnchor) }} className='navbar-link'>Projects</div>
             </div>
             <div className='navbar-link-container'>
               <Link to="/contact" className='navbar-link'>Contact</Link>
@@ -42,7 +42,7 @@ const NavBar = ({showSectionAnchors}) => {
         {/* <div className='navbar-link'>Blog</div> */}
       </div>
       <div className='mobile-nav-container'>
-        <img src={BurgerMenu} onClick={() => setShowDropdown(!showDropdown)}/>
+        <img src={ BurgerMenu } alt='Menu' onClick={() => setShowDropdown(!showDropdown)}/>
         {
           showDropdown ?
             <div className="dropdown-container">
@@ -50,10 +50,10 @@ const NavBar = ({showSectionAnchors}) => {
                 showSectionAnchors ? 
                 <Fragment>
                   <div className='navbar-link-container'>
-                    <div onClick={() => {handleClick(skillsAnchor)}} className='navbar-link'>Skills</div>
+                    <div onClick={() => { handleClick(skillsAnchor) }} className='navbar-link'>Skills</div>
                   </div>
                   <div className='navbar-link-container'>
-                    <div onClick={() => {handleClick(projectAnchor)}} className='navbar-link'>Projects</div>
+                    <div onClick={() => { handleClick(projectAnchor) }} className='navbar-link'>Projects</div>
                   </div>
                   <div className='navbar-link-container'>
                     <Link to="/contact" className='navbar-link'>Contact</Link>
